backend/requirements.txt
fastapi
uvicorn[standard]
redis
pydantic>=2.0
pydantic-settings
python-dotenv
numpy
aiohttp
pandas
requests
classiq
qiskit
matplotlib
scipy
boto3